Some finishing touches on the one-last-sync-before-resetting feature.
I actually experienced a reversal on my opinion of what this feature should do:
Instead of forcing one last sync, then forcing a new sync with the new account, it forces one last sync and resets the account password.
The problem is that clearly your other account is going to have a different password, so there's no need to sync the new account, because it will almost certainly fail. Instead, reset the password so we go in there and set a new password, which causes the app to sync immediately thereafter.





git-svn-id: svn+ssh://taylor-hughes.com/svnroot/idnzo/trunk@197 21c2b8b0-c4d0-461e-9b94-4eb597c105ab